+++ ScatterGather/ScatterGather/Resources/AppDelegate.swift
