+++ ScatterGather/ScatterGather/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
