+++ ScatterGather/ScatterGather/Resources/Assets.xcassets/Contents.json
