+++ ScatterGather/ScatterGather/Resources/Assets.xcassets/lambda_logo.imageset/Contents.json
