+++ ScatterGather/ScatterGather/Resources/Assets.xcassets/lambda_logo.imageset/lambda_logo.png
