+++ ScatterGather/ScatterGather/Resources/Info.plist
