+++ ScatterGather/ScatterGather/Resources/SceneDelegate.swift
